Make the backend and functions private, move tests that need them into lib.rs
